requirements.txt
# Viral Video Finder - 의존성
PyQt5>=5.15.0
APScheduler>=3.10.0
requests>=2.31.0
beautifulsoup4>=4.12.0
selenium>=4.15.0
yt-dlp>=2024.1.0
PyYAML>=6.0
instaloader>=4.10
